feat: configuration OPEN API Client generation solution
