--- src/main/java/info/touret/guitarheaven/infrastructure/ebay/ItemSummary.java
@@ -1,9 +1,0 @@
-package info.touret.guitarheaven.infrastructure.ebay;
-
-
-public record ItemSummary(
-        String itemId,
-        String title,
-        Price price,
-        Image image
-) {}
