--- src/main/java/info/touret/guitarheaven/infrastructure/ebay/Price.java
@@ -1,6 +1,0 @@
-package info.touret.guitarheaven.infrastructure.ebay;
-
-public record Price(
-        double value,
-        String currency
-) {}
