--- src/main/java/info/touret/guitarheaven/infrastructure/ebay/SearchPagedCollection.java
@@ -1,12 +1,0 @@
-package info.touret.guitarheaven.infrastructure.ebay;
-
-
-import java.util.List;
-
-public record SearchPagedCollection(
-        String href,
-        int total,
-        int limit,
-        int offset,
-        List<ItemSummary> itemSummaries
-) {}
