+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # --------------------------
 ## API CLIENT Configuration
 # --------------------------
-quarkus.rest-client."info.touret.guitarheaven.infrastructure.ebay.EbayClient".url=${quarkus.microcks.default.http}/rest/Browse+API/v1.19.9
+quarkus.rest-client.ebay_buy_openapi_yaml.url=${quarkus.microcks.default.http}/rest/Browse+API/v1.19.9
 quarkus.rest-client.extensions-api.verify-host=false
 # --------------------------
 ## KAFKA Client configuration
@@ -35,3 +35,10 @@ mp.openapi.extensions.smallrye.operationIdStrategy=METHOD
 quarkus.microcks.devservices.ensemble.async-enabled=true
 # Customize the Microcks-uber-async-minion image you want to use.
 quarkus.microcks.devservices.ensemble.async-image-name=quay.io/microcks/microcks-uber-async-minion:nightly
+
+
+quarkus.openapi-generator.codegen.input-base-dir=src/main/resources/openapi-client
+quarkus.openapi-generator.codegen.spec.ebay_buy_openapi_yaml.base-package=info.touret.guitarheaven.infrastructure.ebay
+quarkus.openapi-generator.codegen.spec.ebay_buy_openapi_yaml.model-name-suffix=Dto
+quarkus.openapi-generator.codegen.spec.ebay_buy_openapi_yaml.use-bean-validation=true
+
