+++ src/test/resources/application.properties
@@ -1,4 +1,4 @@
-quarkus.rest-client."info.touret.guitarheaven.infrastructure.ebay.EbayClient".url=${quarkus.microcks.default.http}/rest/Browse+API/v1.19.9
+quarkus.rest-client.ebay_buy_openapi_yaml.url=${quarkus.microcks.default.http}/rest/Browse+API/v1.19.9
 quarkus.rest-client.extensions-api.verify-host=false
 quarkus.rest-client.logging.scope=request-response
 quarkus.rest-client.logging.body-limit=50
